add bios info to host report and move os specific info out of shim level
