+++ src/runtime_src/core/common/system.cpp
@@ -2,12 +2,12 @@
 // Copyright (C) 2019 Xilinx, Inc
 // Copyright (C) 2022 Advanced Micro Devices, Inc. All rights reserved.
 #define XRT_CORE_COMMON_SOURCE
+// Local - Include Files
 #include "system.h"
 #include "device.h"
 #include "module_loader.h"
-#include "gen/version.h"
 
-
+// System - Include Files
 #include <vector>
 #include <map>
 #include <memory>
@@ -99,16 +99,9 @@ get_xrt_build_info(boost::property_tree::ptree& pt)
 }
 
 void
-get_xrt_info(boost::property_tree::ptree &pt)
+get_driver_info(boost::property_tree::ptree &pt)
 {
-  get_xrt_build_info(pt);
-  instance().get_xrt_info(pt);
-}
-
-void
-get_os_info(boost::property_tree::ptree& pt)
-{
-  instance().get_os_info(pt);
+  instance().get_driver_info(pt);
 }
 
 void
