+++ src/runtime_src/core/pcie/linux/system_linux.cpp
@@ -1,21 +1,23 @@
 // SPDX-License-Identifier: Apache-2.0
 // Copyright (C) 2019-2022 Xilinx, Inc
 // Copyright (C) 2022 Advanced Micro Devices, Inc. All rights reserved.
+
+// Local - Include files
 #include "device_linux.h"
 #include "system_linux.h"
-#include "gen/version.h"
 #include "pcidev.h"
 #include "pcidrv_xclmgmt.h"
 #include "pcidrv_xocl.h"
-
 #include "core/common/module_loader.h"
 #include "core/common/query_requests.h"
 
+// 3rd Party Library - Include files
 #include <boost/format.hpp>
 #include <boost/property_tree/ini_parser.hpp>
 #include <gnu/libc-version.h>
 #include <sys/utsname.h>
 
+// Local - Include files
 #include <chrono>
 #include <fstream>
 #include <map>
@@ -23,19 +25,6 @@
 #include <thread>
 #include <unistd.h>
 #include <vector>
-
-#if defined(__aarch64__) || defined(__arm__) || defined(__mips__)
-  #define MACHINE_NODE_PATH "/proc/device-tree/model"
-#elif defined(__PPC64__)
-  #define MACHINE_NODE_PATH "/proc/device-tree/model-name"
-  // /proc/device-tree/system-id may be 000000
-  // /proc/device-tree/model may be 00000
-#elif defined (__x86_64__)
-  #define MACHINE_NODE_PATH "/sys/devices/virtual/dmi/id/product_name"
-#else
-#error "Unsupported platform"
-  #define MACHINE_NODE_PATH ""
-#endif
 
 namespace {
 
@@ -97,26 +86,6 @@ driver_version(const std::string& driver)
   _pt.put("version", ver);
   _pt.put("hash", hash);
   return _pt;
-}
-
-static boost::property_tree::ptree
-glibc_info()
-{
-  boost::property_tree::ptree _pt;
-  _pt.put("name", "glibc");
-  _pt.put("version", gnu_get_libc_version());
-  return _pt;
-}
-
-static std::string machine_info()
-{
-  std::string model("unknown");
-  std::ifstream stream(MACHINE_NODE_PATH);
-  if (stream.good()) {
-    std::getline(stream, model);
-    stream.close();
-  }
-  return model;
 }
 
 }
@@ -192,54 +161,13 @@ system_linux()
 
 void
 system_linux::
-get_xrt_info(boost::property_tree::ptree &pt)
+get_driver_info(boost::property_tree::ptree &pt)
 {
   boost::property_tree::ptree _ptDriverInfo;
 
   for (const auto& drv : driver_list::get())
     _ptDriverInfo.push_back( {"", driver_version(drv->name())} );
   pt.put_child("drivers", _ptDriverInfo);
-}
-
-void
-system_linux::
-get_os_info(boost::property_tree::ptree &pt)
-{
-  struct utsname sysinfo;
-  if (!uname(&sysinfo)) {
-    pt.put("sysname",   sysinfo.sysname);
-    pt.put("release",   sysinfo.release);
-    pt.put("version",   sysinfo.version);
-    pt.put("machine",   sysinfo.machine);
-  }
-
-  // The file is a requirement as per latest Linux standards
-  std::ifstream ifs("/etc/os-release");
-  if ( ifs.good() ) {
-      boost::property_tree::ptree opt;
-      boost::property_tree::ini_parser::read_ini(ifs, opt);
-      std::string val = opt.get<std::string>("PRETTY_NAME", "");
-      if (val.length()) {
-          if ((val.front() == '"') && (val.back() == '"')) {
-              val.erase(0, 1);
-              val.erase(val.size()-1);
-          }
-          pt.put("distribution", val);
-      }
-      ifs.close();
-  }
-
-  pt.put("model", machine_info());
-  pt.put("cores", std::thread::hardware_concurrency());
-  pt.put("memory_bytes", (boost::format("0x%lx") % (sysconf(_SC_PHYS_PAGES) * sysconf(_SC_PAGE_SIZE))).str());
-  boost::property_tree::ptree _ptLibInfo;
-  _ptLibInfo.push_back( {"", glibc_info()} );
-  pt.put_child("libraries", _ptLibInfo);
-
-  char hostname[256] = {0};
-  gethostname(hostname, 256);
-  std::string hn(hostname);
-  pt.put("hostname", hn);
 }
 
 device::id_type
