+++ src/runtime_src/core/pcie/windows/alveo/system_windows.h
@@ -12,10 +12,7 @@ class system_windows : public system_pcie
 {
 public:
   void
-  get_xrt_info(boost::property_tree::ptree &pt);
-
-  void
-  get_os_info(boost::property_tree::ptree &pt);
+  get_driver_info(boost::property_tree::ptree &pt);
 
   std::pair<device::id_type, device::id_type>
   get_total_devices(bool is_user) const;
