+++ src/runtime_src/core/tools/xbutil2/CMakeLists.txt
@@ -29,7 +29,6 @@ file(GLOB XBUTIL_V2_BASE_FILES
 
 # Collect local directory files
 file(GLOB XBUTIL_V2_SUBCMD_FILES
-  "XBReport.cpp"
   "SubCmdExamine.cpp"
   "SubCmdProgram.cpp"
   "SubCmdReset.cpp"
