--- src/runtime_src/core/tools/xbutil2/XBReport.h
@@ -1,29 +1,0 @@
-/**
- * Copyright (C) 2019 Xilinx, Inc
- *
- * Licensed under the Apache License, Version 2.0 (the "License"). You may
- * not use this file except in compliance with the License. A copy of the
- * License is located at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
- * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
- * License for the specific language governing permissions and limitations
- * under the License.
- */
-
-#ifndef __XBReport_h_
-#define __XBReport_h_
-
-// Include files
-// Please keep these to the bare minimum
-
-namespace XBReport {
-  void report_system_config();
-  void report_xrt_info();
-};
-
-#endif
-
